Move EventDetectionManager.py and AnomalyDetectionManager.py to /DetectionManagers
