+++ DetectionManagers/AnomalyDetectionManager.py
